+++ TgDownloaderConsole/Helpers/MenuHelperDownload.cs
@@ -1,6 +1,7 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
-// See https://aka.ms/new-console-template for more information
+
+using TgStorageCore.Models.SourcesSettings;
 
 namespace TgDownloaderConsole.Helpers;
 
@@ -20,20 +21,26 @@ internal partial class MenuHelper
                     TgLocale.MenuDownloadSetSourceId, 
                     TgLocale.MenuDownloadSetSourceUserName, 
                     TgLocale.MenuDownloadSetFolder, 
+                    TgLocale.MenuDownloadSetSourceStartIdAuto, 
+                    TgLocale.MenuDownloadSetSourceStartIdManual, 
                     TgLocale.MenuDownloadSetIsRewriteFiles,
                     TgLocale.MenuDownloadSetIsRewriteMessages,
                     TgLocale.MenuDownloadSetIsAddMessageId,
                     TgLocale.MenuDownload
+                    //TgLocale.MenuScanMyChannels
                 ));
         return userChoose switch
         {
             "Setup source ID" => MenuDownload.SetSourceId,
             "Setup source user name" => MenuDownload.SetSourceUserName,
             "Setup download folder" => MenuDownload.SetDestDirectory,
+            "Setup source start ID auto" => MenuDownload.SetSourceStartIdAuto,
+            "Setup source start ID manual" => MenuDownload.SetSourceStartIdManual,
             "Enable rewrite exists files" => MenuDownload.SetIsRewriteFiles,
             "Enable rewrite exists messages" => MenuDownload.SetIsRewriteMessages,
             "Enable join message ID with file name" => MenuDownload.SetIsAddMessageId,
             "Download" => MenuDownload.Download,
+            //"Scan my channels and groups" => MenuDownload.ScanChannels,
             _ => MenuDownload.Return
         };
     }
@@ -48,27 +55,45 @@ internal partial class MenuHelper
             switch (menu)
             {
                 case MenuDownload.SetSourceId:
-                    TgClient.TgDownload.SetSourceIdByAsk();
+                    SetTgDownloadSourceId();
                     TgClient.PrepareDownloadMessages(true);
+                    LoadTgClientSettings();
+                    SetSourceWithSettings();
+                    break;
+                case MenuDownload.SetSourceStartIdAuto:
+                    RunAction(SetTgDownloadSourceStartIdAuto, true);
+                    LoadTgClientSettings();
+                    SetSourceWithSettings();
+                    break;
+                case MenuDownload.SetSourceStartIdManual:
+                    SetTgDownloadSourceStartIdManual();
+                    LoadTgClientSettings();
+                    SetSourceWithSettings();
                     break;
                 case MenuDownload.SetSourceUserName:
-                    TgClient.TgDownload.SetSourceUserNameByAsk();
+                    SetTgDownloadSourceUserName();
                     TgClient.PrepareDownloadMessages(true);
+                    LoadTgClientSettings();
+                    SetSourceWithSettings();
                     break;
                 case MenuDownload.SetDestDirectory:
-                    TgClient.TgDownload.SetDestDirectory();
+                    SetTgDownloadDestDirectory();
+                    SetSourceWithSettings();
                     break;
                 case MenuDownload.SetIsRewriteFiles:
-                    TgClient.TgDownload.SetIsRewriteFiles();
+                    SetTgDownloadIsRewriteFiles();
                     break;
                 case MenuDownload.SetIsRewriteMessages:
-                    TgClient.TgDownload.SetIsRewriteMessages();
+                    SetTgDownloadIsRewriteMessages();
                     break;
                 case MenuDownload.SetIsAddMessageId:
-                    TgClient.TgDownload.SetIsAddMessageId();
+                    SetTgDownloadIsJoinFileNameWithMessageId();
                     break;
                 case MenuDownload.Download:
-                    RunAction(Download);
+                    RunAction(Download, false);
+                    break;
+                case MenuDownload.ScanChannels:
+                    RunAction(ScanRange, true);
                     break;
                 case MenuDownload.Return:
                 default:
@@ -77,11 +102,131 @@ internal partial class MenuHelper
         } while (menu is not MenuDownload.Return);
     }
 
-    public void Download()
+    private void SetTgDownloadSourceId()
+    {
+        TgClient.TgDownload.SetDefault(1);
+        bool isCheck;
+        long sourceId;
+        do
+        {
+            sourceId = AnsiConsole.Ask<long>(TgLog.GetLineStampInfo(TgLocale.TypeTgSourceId));
+            isCheck = TgClient.TgDownload.IsReadySourceId;
+        } while (!isCheck);
+        TgClient.TgDownload.SourceId = sourceId;
+    }
+
+    private void SetTgDownloadSourceStartIdAuto()
+    {
+        Channel? channel = TgClient.PrepareDownloadMessages(true);
+        if (channel is null) return;
+        TgClient.SetChannelMessageIdFirst(channel, RefreshStatusForDownload);
+    }
+
+    private void SetTgDownloadSourceStartIdManual()
+    {
+        bool isCheck;
+        int startId;
+        do
+        {
+            startId = AnsiConsole.Ask<int>(TgLog.GetLineStampInfo(TgLocale.TypeTgSourceStartId));
+            isCheck = TgClient.TgDownload.IsReadySourceStartId;
+        } while (!isCheck);
+        TgClient.TgDownload.SourceStartId = startId;
+    }
+
+    private void SetTgDownloadSourceUserName()
+    {
+        TgClient.TgDownload.SetDefault(1);
+        bool isCheck;
+        string sourceUserName;
+        do
+        {
+            sourceUserName = AnsiConsole.Ask<string>(TgLog.GetLineStampInfo(TgLocale.TypeTgSourceUserName));
+            if (!string.IsNullOrEmpty(sourceUserName))
+            {
+                sourceUserName = sourceUserName.StartsWith(@"https://t.me/")
+                    ? sourceUserName.Replace("https://t.me/", string.Empty)
+                    : sourceUserName;
+            }
+
+            isCheck = !string.IsNullOrEmpty(sourceUserName);
+        } while (!isCheck);
+        TgClient.TgDownload.SourceUserName = sourceUserName;
+    }
+
+    private void SetTgDownloadDestDirectory()
+    {
+        string destDirectory;
+        do
+        {
+            destDirectory = AnsiConsole.Ask<string>(TgLog.GetLineStampInfo(TgLocale.TypeDestDirectory));
+            if (!Directory.Exists(destDirectory))
+                TgLog.Info(TgLocale.DirIsNotExistsSpecify(destDirectory));
+        } while (!Directory.Exists(destDirectory));
+        TgClient.TgDownload.DestDirectory = destDirectory;
+    }
+
+    private void SetTgDownloadIsRewriteFiles()
+    {
+        bool isResult = AnsiConsole.Prompt(new SelectionPrompt<bool>()
+            .Title(TgLocale.TgSettingsIsRewriteFiles)
+            .PageSize(10)
+            //.MoreChoicesText("[grey](Move up and down to reveal more)[/]")
+            .AddChoices(true, false));
+        TgClient.TgDownload.IsRewriteFiles = isResult;
+    }
+
+    private void SetTgDownloadIsRewriteMessages()
+    {
+        bool isResult = AnsiConsole.Prompt(new SelectionPrompt<bool>()
+            .Title(TgLocale.TgSettingsIsRewriteMessages)
+            .PageSize(10)
+            //.MoreChoicesText("[grey](Move up and down to reveal more)[/]")
+            .AddChoices(true, false));
+        TgClient.TgDownload.IsRewriteMessages = isResult;
+    }
+
+    private void SetTgDownloadIsJoinFileNameWithMessageId()
+    {
+        bool isResult = AnsiConsole.Prompt(new SelectionPrompt<bool>()
+            .Title(TgLocale.TgSettingsIsJoinFileNameWithMessageId)
+            .PageSize(10)
+            //.MoreChoicesText("[grey](Move up and down to reveal more)[/]")
+            .AddChoices(true, false));
+        TgClient.TgDownload.IsJoinFileNameWithMessageId = isResult;
+    }
+
+    private void SetSourceWithSettings()
+    {
+        if (!TgClient.TgDownload.IsReady) return;
+        TgStorage.AddOrUpdateRecordSource(TgClient.TgDownload.SourceId, TgClient.TgDownload.SourceUserName, 
+            TgClient.TgDownload.SourceTitle, TgClient.TgDownload.SourceAbout, TgClient.TgDownload.SourceLastId, true);
+        TgStorage.AddOrUpdateRecordSourceSetting(TgClient.TgDownload.SourceId, TgClient.TgDownload.DestDirectory, true);
+    }
+
+    private void LoadTgClientSettings()
+    {
+        TableSourceSettingModel sourceSettings = TgStorage.GetRecord<TableSourceSettingModel>(null, TgClient.TgDownload.SourceId);
+        if (!TgClient.TgDownload.IsReadyDescription)
+            TgClient.TgDownload.DestDirectory = sourceSettings.Directory;
+    }
+
+    private void StoreSource(ChatBase chat, string about, int count)
+    {
+        if (chat is Channel channel)
+            TgStorage.AddOrUpdateRecordSource(channel.id, channel.username, channel.title, about, count, true);
+    }
+
+    private void Download()
     {
         ShowTableDownload();
-        TgStorage.AddOrUpdateRecordSource(TgClient.TgDownload.SourceId, TgClient.TgDownload.SourceUserName, false);
         TgClient.DownloadAllData(RefreshStatusForDownload, StoreMessage, StoreDocument, FindExistsMessage);
+    }
+
+    private void ScanRange()
+    {
+        ShowTableScanRange();
+        TgClient.FindAndStoreChannel(RefreshStatusForDownload, StoreSource);
     }
 
     #endregion